# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SPARK AI Marketing Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in the organization
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: inviteUser
      tags: [users]
      summary: Invite a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInviteInput"
      responses:
        "201":
          description: User invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: removeUser
      tags: [users]
      summary: Remove user from organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, organizationId, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, marketing_manager, channel_specialist, creative_reviewer, finance_approver, viewer]
        organizationId:
          type: integer
        organizationName:
          type: ["string", "null"]
        organizationPlan:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, pending]
        createdAt:
          type: string
    UserInviteInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
    UserUpdate:
      type: object
      properties:
        role:
          type: string
        status:
          type: string

    # ── Dashboard ──
...